app/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
prometheus-fastapi-instrumentator==6.1.0
